Updated folder structure for Assignment-1-Tests
