+++ Assignment-4/Assignment-1-Tests/LibraryManagerTests.cs
@@ -1,13 +1,15 @@
-﻿using assignment_1.library.books;
+﻿using Assignment_1_Tests.TestData;
 using assignment_1.library.libraryManagement;
+using assignment_1.library.books;
 using assignment_1.library.users;
+using Xunit;
 
 namespace Assignment_1_Tests
 {
     public class LibraryManagerTests
     {
         [Theory]
-        [MemberData(nameof(TestIssueBookData))]
+        [MemberData(nameof(TestCalculateFineData.TestIssueBookData), MemberType = typeof(TestCalculateFineData))]
         //It includes different combinations of users, books, and expected copies available after issuing
         public void TestIssueBook(User user, Book book, int expectedCopiesAvailable)
         {
@@ -23,15 +25,8 @@ namespace Assignment_1_Tests
             Assert.Equal(expectedCopiesAvailable, book.CopiesAvailable);
         }
 
-        public static IEnumerable<object[]> TestIssueBookData()
-        {
-            yield return new object[] { new Student("id-1", "Prashant Shukla", "st-1"), new Book("b-1", "Intro C#", "Author-1", 5), 4 };
-            yield return new object[] { new Teacher("id-2", "Piyush", "t-1", "Maths"), new Book("b-2", "DSA Book", "Author-2", 10), 9 };
-            yield return new object[] { new Admin("id-3", "Admin User", "admin-1", "Administrator"), new Book("b-3", "DevOps Book", "Author-3", 3), 2 };
-        }
-
         [Theory]
-        [MemberData(nameof(TestReturnBookData))]
+        [MemberData(nameof(TestCalculateFineData.TestReturnBookData), MemberType = typeof(TestCalculateFineData))]
         //Verifies whether the number of copies available is updated correctly after returning a book
         public void TestReturnBook(User user, Book book, int expectedCopiesAvailable)
         {
@@ -52,15 +47,8 @@ namespace Assignment_1_Tests
             Assert.DoesNotContain(book, user.BooksIssued);
         }
 
-        public static IEnumerable<object[]> TestReturnBookData()
-        {
-            yield return new object[] { new Student("id-1", "Prashant Shukla", "st-1"), new Book("b-1", "Intro C#", "Author-1", 5), 5 };
-            yield return new object[] { new Teacher("id-2", "Piyush", "t-1", "Maths"), new Book("b-2", "DSA Book", "Author-2", 10), 10 };
-            yield return new object[] { new Admin("id-3", "Admin User", "admin-1", "Administrator"), new Book("b-3", "DevOps Book", "Author-3", 3), 3 };
-        }
-
         [Theory]
-        [MemberData(nameof(TestIssueBook_NotEnoughCopiesData))]
+        [MemberData(nameof(TestCalculateFineData.TestIssueBook_NotEnoughCopiesData), MemberType = typeof(TestCalculateFineData))]
         //Verifies whether the number of copies available remains unchanged(still 0) after attempting to issue the book
         public void TestIssueBook_NotEnoughCopies(User user, Book book)
         {
@@ -76,14 +64,5 @@ namespace Assignment_1_Tests
             // CopiesAvailable should remain unchanged (still 0)
             Assert.Equal(0, book.CopiesAvailable);
         }
-
-        public static IEnumerable<object[]> TestIssueBook_NotEnoughCopiesData()
-        {
-            yield return new object[] { new Student("id-1", "Prashant Shukla", "st-1"), new Book("b-1", "Intro C#", "Author-1", 0) };
-            yield return new object[] { new Teacher("id-2", "Piyush", "t-1", "Maths"), new Book("b-2", "DSA Book", "Author-2", 0) };
-            yield return new object[] { new Admin("id-3", "Admin User", "admin-1", "Administrator"), new Book("b-3", "DevOps Book", "Author-3", 0) };
-        }
-
     }
 }
-
